+++ MakeMyCapServer/Shopify/Dtos/Fulfillment/LineItem.cs
@@ -6,7 +6,7 @@ namespace MakeMyCapServer.Shopify.Dtos.Fulfillment;
     public class LineItem
     {
         [JsonPropertyName("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonPropertyName("admin_graphql_api_id")]
         public string AdminGraphqlApiId { get; set; }
@@ -39,7 +39,7 @@ namespace MakeMyCapServer.Shopify.Dtos.Fulfillment;
         public bool ProductExists { get; set; }
 
         [JsonPropertyName("product_id")]
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         [JsonPropertyName("properties")]
         public List<Property> Properties { get; set; }
@@ -66,7 +66,7 @@ namespace MakeMyCapServer.Shopify.Dtos.Fulfillment;
         public AmountSet TotalDiscountSet { get; set; }
 
         [JsonPropertyName("variant_id")]
-        public int VariantId { get; set; }
+        public long VariantId { get; set; }
 
         [JsonPropertyName("variant_inventory_management")]
         public string VariantInventoryManagement { get; set; }
